backend/pyproject.toml
[project]
name = "cardiosphere-backend"
version = "1.0.0"
description = "CardioSphere — Heart Disease Risk Prediction API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "motor==3.3.2",
    "pymongo==4.6.1",
    "pydantic==2.5.2",
    "pydantic-settings==2.1.0",
    "python-jose[cryptography]==3.3.0",
    "PyJWT[crypto]==2.8.0",
    "cryptography==41.0.7",
    "httpx==0.25.2",
    "python-multipart==0.0.6",
    "joblib==1.3.2",
    "scikit-learn==1.3.2",
    "pandas==2.1.4",
    "numpy==1.26.2",
    "openai==1.6.1",
    "Pillow==10.1.0",
    "python-dotenv==1.0.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.1.2",
    "aiofiles==23.2.1",
    "certifi>=2024.0.0",
    "twilio>=9.10.2",
    "apscheduler>=3.11.2",
    "tzlocal>=5.3.1",
    "pytz>=2025.2",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.25.0",
]

[tool.uv]
link-mode = "copy"
